Se ajustan endpoints para creación de información cursos, docentes, estudiantes, categorias

+++ src/main/java/co/proyectoGrado/proyectoGrado/domain/model/CursoContenido.java
@@ -9,8 +9,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CursoContenido {
+
     @JsonProperty
-    private int idCursoContenido;
+    private Long idCursoContenido;
     @JsonProperty
     private int idCategoriaContenido;
     @JsonProperty
